// generated/src/app/api-models/vendor-adv-pay-acc-detail.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class VendorAdvPayAccDetail {
  seqVendAdvPayAccDtl: number;
  seqVendAdvPayAcc: number;
  sequenceNumber: number;
  advPayAmount: number;
  checkEftType: string;
  checkEftNumber: string;
  checkEftDate: Date;
  checkEftMethod: string;
  apStatus: string;
  balanceAmount: number;
  selectedForPay: number;
  seqApTrans: number;
  approveStatus: string;
  approveUserId: string;
  approveDate: Date;
  denyReason: string;
  appUserDefined1: string;
  appUserDefined2: string;
  appUserDate1: Date;
  appUserDate2: Date;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}